Changed libraries

+++ libft/ft_putnbr_fd.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putnbr_fd.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: qmanamel <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: ttshivhu <marvin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2017/06/01 12:46:41 by qmanamel          #+#    #+#             */
-/*   Updated: 2017/06/01 12:46:42 by qmanamel         ###   ########.fr       */
+/*   Created: 2017/05/24 18:51:53 by ttshivhu          #+#    #+#             */
+/*   Updated: 2017/08/04 09:28:20 by ttshivhu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ static void		min_int(int sign, int fd)
 		ft_putstr_fd("2147483647", fd);
 }
 
-void			ft_putnbr_fd(int n, int fd)
+void			ft_putnbr_fd(long long n, int fd)
 {
 	if (n == -2147483648)
 		return (min_int(1, fd));
